test: split PDF prose hyphen continuity cases
